core: add web-safe file and trace services
